Refactor and optimize file handling functions; update header inclusions and improve memory management for person records

+++ EDIII_Trabalho1/io_cabecalho.h
@@ -5,8 +5,8 @@
 #include "data_structs.h"
 
 void escreveCabecalhoIndice(FILE *arquivoIndice, char status);
-void escreveCabecalhoPessoa(FILE *arquivoPessoa, char status, int quantidadePessoas, int quantidadeRemovidos, long long int proxByteOffset);
-void escreveCabecalhoSegue(FILE *arquivoSegue, char status, int quantidadePessoas, long long int proxRRN);
+void escreveCabecalhoPessoa(FILE *arquivoPessoa, char status, int quantidadePessoas, int quantidadeRemovidos, long proxByteOffset);
+void escreveCabecalhoSegue(FILE *arquivoSegue, char status, int quantidadePessoas, int proxRRN);
 CabecalhoSegue leCabecalhoSegue(FILE *arquivoSegue);
 CabecalhoPessoa leCabecalhoPessoa(FILE *arquivoPessoa);
